llm_core/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-doc
annotated-types
anyio
asyncio
attrs
certifi
charset-normalizer
click
diskcache
distro
fastapi
filelock
frozenlist
fsspec
grpcio
h11
h2
hf-xet
hpack
httpcore
httptools
httpx
huggingface-hub
hyperframe
idna
Jinja2
jiter
joblib
loguru
MarkupSafe
mpmath
multidict
networkx
numpy
openai
packaging
portalocker
propcache
protobuf
pydantic
pydantic-settings
pydantic_core
python-dotenv
PyYAML
qdrant-client
regex
requests
safetensors
sentence-transformers
setuptools
sniffio
starlette
sympy
threadpoolctl
tokenizers
tqdm
transformers
typing-inspection
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websockets
yarl
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-semantic-conventions==0.59b0
